scripts: footprint: fix node size computation
fix issue with (no paths) != sum(<no paths>)

In some cases it was observed that item were being
added to the node_no_paths and these only had a
path of ':' hence no node was ever added.

This resulted in the size of (no paths) to not be
equal to the sum of it's children.

I am not sure if this size should be a child named
of no paths or if  this should be part of the
node_hidden_syms. I assumed it should be part
of node_hidden_syms

(cherry picked from commit 003fb88afdead00a40dab4503c69f061ecf6ad4d)

Original-Signed-off-by: Kyle Micallef Bonnici <kylebonnici@hotmail.com>
GitOrigin-RevId: 003fb88afdead00a40dab4503c69f061ecf6ad4d
Cr-Build-Id: 8715487325654683073
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8715487325654683073
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ibb4dc050ed43d9ac9e5af94e761ac89d9a522364
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6521806
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>