drivers: serial: cc23x0: Add support for DMA mode
Two DMA channels are assigned to TX and RX respectively:
- A TX DMA request is asserted when there is space in the FIFO.
- A RX DMA request is asserted when data is in the FIFO.

When DMA is enabled for a peripheral, the DMA transfer completion is
signaled on the peripheral's interrupt only (here UART's interrupt).
It is not signaled on the DMA dedicated interrupt.

Also, when DMA is enabled for a peripheral, the DMA controller stops
the normal transfer interrupts for this peripheral from reaching the
NVIC (the interrupts are still reported in the interrupt registers of
the peripheral). Thus, when a large amount of data is transferred using
DMA, instead of receiving multiple interrupts from the peripheral as
data flows, the NVIC receives only one interrupt when the transfer
completes (unmasked peripheral error interrupts continue to be sent
to the NVIC).

(cherry picked from commit fa199b6f6cfdf19444ab7c68138f2038b62f6c8d)

Original-Signed-off-by: Julien Panis <jpanis@baylibre.com>
GitOrigin-RevId: fa199b6f6cfdf19444ab7c68138f2038b62f6c8d
Cr-Build-Id: 8712452327885524593
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8712452327885524593
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0d05d090eee39542c1a6332f3cbdaababd7fe508
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6633191
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>