samples: net: openthread: Add a sample using OpenThread CoAP API
In order to show how to use OpenThread and CoAP, add an application.
This example could be build to run a client or server.
The server could expose LEDs and buttons and client could get their
state and set the LEDs state.

The network is created automatically using the network key predefined
in the config. The goal is to make the example simple by removing
the commisionning and joinning process.

If a client application has a button, it could use it to toggle the first
LED on any boards running the server application.

(cherry picked from commit dc898b47c83151d68478708c36a7a20259706be7)

Original-Signed-off-by: Alexandre Bailon <abailon@baylibre.com>
GitOrigin-RevId: dc898b47c83151d68478708c36a7a20259706be7
Cr-Build-Id: 8725090605295300081
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8725090605295300081
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Icb3fb13957dd22d6daf1e9515073f10cbc7cd6e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6188406
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>