usb: device_next: allow CDC ACM UART to be initialized and enabled at boot
There are some boards and samples in the tree that use the CDC ACM UART
as the default serial backend, just like a real UART controller would.
The new device stack requires more detailed configuration than the old
one. In order to use the CDC ACM UART as the serial backend with the new
device stack in the same way as with the legacy stack, we need to
provide a solution to initialise and enable the CDC ACM UART at boot
time. We cannot use snippets as they do not support Kconfig files or
source code. Shields would be an option, but they cannot be used for
virtual devices such as the CDC ACM UART. The remaining solution is to
put the code and Kconfig file for it in the subsys directory.

Allow CDC ACM UART instance and USB device stack to be initialized and
enabled at boot time and to use it as serial backend for logging or
shell.

(cherry picked from commit 7ffa6203e1154cbd8d407ee53516914ad504b7eb)

Original-Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>
GitOrigin-RevId: 7ffa6203e1154cbd8d407ee53516914ad504b7eb
Cr-Build-Id: 8723142769276654049
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8723142769276654049
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7675a7c98ef193c0b5199cad3bafa1f4016916e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6257268
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>