tests: watchdog: Add missing wdt_disable in API test
The current watchdog tests assume that watchdog can be configured once
it is running. This is a concept that do not work in practice becasue
most of watchdog once configured will be locked. The tests should take
that in consideration before enable another test. The other important
details is that even when a watchdog is running it should be disable in
some controllers to allow reconfigure.

This change was tested with SAM4L watchdog controller #83475 which
allows reconfigure the watchdog once it is enabled. To do the correct
configuration it is necessary execute the wdt_disable() call.

The change will validate if the watchdog can be disabled for those
cases were the controller is not locked and fallback the test on the
contrary.

(cherry picked from commit 0c4f1987c6cc5c69d9362a4ffaa08403ebb1be81)

Original-Signed-off-by: Gerson Fernando Budke <nandojve@gmail.com>
GitOrigin-RevId: 0c4f1987c6cc5c69d9362a4ffaa08403ebb1be81
Cr-Build-Id: 8723550454920298305
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8723550454920298305
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I914bc71ce40eeddd9e2ae0f082afa803dc15e24d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6243012
Commit-Queue: Jonathon Murphy <jpmurphy@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>