llext-edk: import data from build_info.yml and .config
Currently, the llext-edk.cmake script requires a number of variables to
be passed in from the main CMakeLists.txt file as arguments to be able
to customize the generated files.

To improve this rigid approach, the script is modified to read in the
following files in the build directory:

 * 'zephyr/.config', for the final set of Kconfig options used;
 * 'build_info.yml', for the cmake-related variables.

This is more flexible and also easier to maintain, as it doesn't require
manual changes to the main CMakelists.txt file when new variables need
to be referenced.

(cherry picked from commit 5d3fe7e85a0d569eada39bbdadc388835646737e)

Original-Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
GitOrigin-RevId: 5d3fe7e85a0d569eada39bbdadc388835646737e
Cr-Build-Id: 8723550454920298305
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8723550454920298305
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I48b93438a44dfa4cd1a11bcf54a24db2e86d0fc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6245693
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Jonathon Murphy <jpmurphy@google.com>