llext: ignore relocations on non-allocated sections
When building an ELF file with debugging information, the compiler
places this data in sections that do not need to be loaded in memory for
the program to run. These sections are marked as "not allocated" in the
ELF section header, and can be safely ignored when loading the file.

This patch adds a check to the relocation processing code to skip
relocations acting on those non-allocated sections, allowing to load ELF
files with debugging information.

(cherry picked from commit cf11ab4b0fa2663dac2083e5a6d5e6f3cce4dc19)

Original-Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
GitOrigin-RevId: cf11ab4b0fa2663dac2083e5a6d5e6f3cce4dc19
Cr-Build-Id: 8724592320935143761
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8724592320935143761
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ie99e9963a410a38c5c6a596df175b90442c40db8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6201523
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>