kernel: sys_heap: stats: save heap pointers to an array during init
To request heap statistics, a pointer to a heap structure is required.
This is straightforward for a user-defined heap. However, such a pointer
is not known for heaps created by other components or libraries, like
libc. Therefore, it is not possible to calculate the total heap memory.

The proposed solution is to use an array of pointers, which is filled in
on every sys_heap_init() call. One can then iterate through it to sum up
the total memory allocated for all heaps.

The array size is configurable. The default array size is zero,
which means the feature is disabled. Any other integer greater then zero
defines the array size and enables the feature.

A list of pointers instead of an array could be another approach,
but it requeres a heap, which is not always available.

(cherry picked from commit 257d9d45ba0ae8fb08f44faff18a72bd48682081)

Original-Signed-off-by: Ivan Pankratov <ivan.pankratov@silabs.com>
GitOrigin-RevId: 257d9d45ba0ae8fb08f44faff18a72bd48682081
Cr-Build-Id: 8724909411128659441
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8724909411128659441
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I460434263d94d315874268d49fdcda726bcaf48a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6196377
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>