drivers: spi_context: Functions for remaining len
Add functions to get remaining length in the transfer. Refactor a bit to
avoid duplicate code for the for loop that is the same as in the total
length function.

The difference between the total length and left length function is that
the current buffer total length is counted in the former and the current
buffer remaining length is counted in the latter.

(cherry picked from commit 0bb7ccbb0c0fe849d30096ed53e329d3c1cb8307)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 0bb7ccbb0c0fe849d30096ed53e329d3c1cb8307
Cr-Build-Id: 8724275231234053617
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8724275231234053617
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7090638b0560be77a0104782135e201c23b1492a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6220353
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>