net: http_server: fix URL matching with '?' character in resource
Fixes #84198.

If a '?' character is used as part of a wildcard resource, do not treat
this as the end of the string when comparing with a path from the HTTP
request. Only the path from the HTTP request may be terminated by '?'
(in the case of a request with query parameters).

(cherry picked from commit d1d85fa40b3071550361c7520daf90ad273e6b35)

Original-Signed-off-by: Matt Rodgers <mrodgers@witekio.com>
GitOrigin-RevId: d1d85fa40b3071550361c7520daf90ad273e6b35
Cr-Build-Id: 8725000003341292065
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8725000003341292065
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ia8b4c0e3af201d146292e7ed32cf9a06e30650e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6192698
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>