tests: bsim: cap_broadcast_reception: Workaround time dependent issue
This test is overly sensitive to the relative timing of the devices
and/or their random number generation.
Due to this the test failed with its current setup when chanding the
random generation for the nrf54.
Let's work around it by offsetting a bit one of the devices.
This is a provisional workaround until the underlying issue is
addressed.

(cherry picked from commit d961258788e7d9be8b69af9a33935f8099aa4a72)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: d961258788e7d9be8b69af9a33935f8099aa4a72
Cr-Build-Id: 8725000003341292065
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8725000003341292065
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7ff66db2f2e37606e7dff5fb19849d862c2a8064
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6192715
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>