doc: the west workspace topdir is not a git repository
It's a design decision that the west workspace "topdir"
(that is, the top level directory containing the .west directory), is
not itself a git repository. This exists to give us some breathing
room to make changes to the workspace which would otherwise
potentially cause breakage if the entire workspace is in .git.

While this has always been the case, I'm documenting this now because
I reviewed a PR today that flipped my bit from "this is a question to
answer on a case by case basis" to "this is a frequently enough asked
question that I want to be able to link people to the answer":

https://github.com/zephyrproject-rtos/zephyr/pull/84305

We can debate the wisdom of this design decision (and, informally,
I've always tried to avoid breaking these setups), but IMO it's too
late to try to change this in west. Make it explicit in the docs that
you're on your own if you try this.

(cherry picked from commit 7d40091fbfedfc51596447212b408a06cefd702b)

Original-Signed-off-by: Martí Bolívar <marti.bolivar@oss.qualcomm.com>
GitOrigin-RevId: 7d40091fbfedfc51596447212b408a06cefd702b
Cr-Build-Id: 8724909411128659441
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8724909411128659441
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I215b854a8369a09c23d00130edfab80cc1c983fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6196379
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>