doc boards nrfbsim: Update list of supported HW
The models include enough of the CRACEN to run the same
nrfx drivers as in Zephyr. So let's add it to the list.

Also let's remove the RTC from the list as the hal does not
expose it anymore for this platform as the GRTC is to be used
instead.

(cherry picked from commit a211999141833bba7dfd3336f8e3b9b135bb4627)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: a211999141833bba7dfd3336f8e3b9b135bb4627
Cr-Build-Id: 8724909411128659441
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8724909411128659441
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I84f5ff48c4cea7ad4b2715aa529c250339021582
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6196372
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>