tests: drivers: spi: loopback: Add test for pico-spi-pio driver
Plus refactored the structure of the overlay files.
The wiring of the new test is the same as the existing spi tests
for rpi-pico.

The speed of the fast test has to be reduced to ensure signal integrity.

Setting CONFIG_SPI_ASYNC will trigger a hard fault. Seems to be a bug in
the driver.

(cherry picked from commit cef0c1af0d8f6c53813d3f63861235d917825617)

Original-Signed-off-by: Terry Geng <terry@terriex.com>
GitOrigin-RevId: cef0c1af0d8f6c53813d3f63861235d917825617
Cr-Build-Id: 8724411126655316049
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8724411126655316049
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Icd22c10108543ac5d9af39aef2066c252a7cbbc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6212749
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>