lib: os: mpsc_pbuf: Fix infinite loop during allocation
It was possible that allocation was continuously attempting to find
space. It could happen if allocation interrupted consuming a packet
(which is marked as busy) and there is not enough space to allocate
requested packet but there would be if busy packet was freed.
Algorithm in that case was continuously going through the buffer
in search for packets that can be dropped to find space for the
new packet.

(cherry picked from commit 3e52109590cc6d3e226d30587bfabe120a4a4db8)

Original-Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>
GitOrigin-RevId: 3e52109590cc6d3e226d30587bfabe120a4a4db8
Cr-Build-Id: 8724365828692706609
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8724365828692706609
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Iffc707ceb7a3ee3470db8da998057927282532bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6217021
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>