boards: nrf_bsim: Document how to use the console with bsim
This commit describes how to build and run applications using
the console. This documentation is added because it may
not be obvious how this is done. That is, sometimes a user
may believe that the terminal launching the application is the
the console which is not the case.

(cherry picked from commit 0f3e6fd3d624f3647005b8f2bc30bd477933a60a)

Original-Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>
GitOrigin-RevId: 0f3e6fd3d624f3647005b8f2bc30bd477933a60a
Cr-Build-Id: 8723550454920298305
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8723550454920298305
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I18f49d216998b57f878717e0b21ece4fa80c0184
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6244119
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>
Commit-Queue: Jonathon Murphy <jpmurphy@google.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>