net: l2: Add a config option to allow mismatched L3/L2 address packets
The RFC1122 section 3.3.6 says we SHOULD drop the packets if L2 address
is brodcast but L3 address is unicast, but we had seen some Wi-Fi access
points in the field not conforming to that, and DHCP offer is dropped
due to this and causes Wi-Fi connectivity issues.

As the RFC saus it's SHOULD and not a MUST, add a config option to allow
such packets, disabled by default.

(cherry picked from commit 4cd6654b2196bf00d62a528707e45c15db8b627d)

Original-Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>
GitOrigin-RevId: 4cd6654b2196bf00d62a528707e45c15db8b627d
Cr-Build-Id: 8723641050355720609
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8723641050355720609
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0166e9e3a6bd753cbc2b80e8918ccd8456c34c08
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6241418
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>