net: lwm2m: Fix offset for post-write callback
The offset in post-write callback is incorrect for TLV format as
the first data block contain TLV header. This fix is replacing
reamaning for offset in opaque context struct and passing it to
post-write and validation callback.

The fix also corrects the offset in pull context as it was pointing
at the end of the block processed data and not on the block context
prior to processing.

(cherry picked from commit 31ec42a33c08546313dd278956e214550c877523)

Original-Signed-off-by: Dave Lacerte <lacerte.dave@hydroquebec.com>
GitOrigin-RevId: 31ec42a33c08546313dd278956e214550c877523
Cr-Build-Id: 8723731651065188129
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8723731651065188129
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I1f1ef515bbacb7ab314d9e28fda08a3c70346d40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6237949
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>