pm: policy: Fix possible undefined reference
When PM_POLICY_DEVICE_CONSTRAINTS is enabled, we create a list of
devices with the property “zephyr,disabling-power-states” and their
respective states to lock. However, if a device is declared but not
built (e.g., due to a Kconfig option), we have an undefined reference.

This PR changes the code to use the device name instead of the
reference. This increases lookup time (which could be cached at
runtime), and we still have an entry for a non-existent device.

(cherry picked from commit b8ce1761095e6916bbb342f74024cd4342e33334)

Original-Signed-off-by: Flavio Ceolin <flavio@hubblenetwork.com>
GitOrigin-RevId: b8ce1761095e6916bbb342f74024cd4342e33334
Cr-Build-Id: 8723505156996810497
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8723505156996810497
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Id29e8af2d8c75307afe1a30694b0803c56eb5413
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6246483
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>