net: dns: Allow creating context with specific interface
Allow user to specify the network interface where the query should
be sent, but only if the user has supplied the DNS servers in sockaddr
list. If user has specified a DNS servers in a string list, then the
possible network interface is taken from the string.

The reason for this is that it is not possible to add network interface
to the sockaddr list so an extra parameter is needed.

(cherry picked from commit 8d601a0cc261e7c4ab9d5d99b666c427a65333f2)

Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: 8d601a0cc261e7c4ab9d5d99b666c427a65333f2
Cr-Build-Id: 8723142769276654049
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8723142769276654049
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0b9776d21456f3752bcd78b4e62281f627cad357
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6257395
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>