llext-edk: merge target-specific file generation
This commit removes all CMake- or Makefile-specific variables from the
code and instead uses a single list for each type of flag. Where a
difference is needed between the two, a special marker is used.
These markers are replaced later in the target-specific area of the
write functions.

Escaping of problematic characters is also added to the write functions,
to ensure that quotes and backslashes in the flags are properly handled.

(cherry picked from commit 701f5797fc2183743595ef550619aee8f42e436e)

Original-Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
GitOrigin-RevId: 701f5797fc2183743595ef550619aee8f42e436e
Cr-Build-Id: 8723550454920298305
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8723550454920298305
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic60920abd211d4b51eea5d72dff4f63643d4c0c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6245692
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Jonathon Murphy <jpmurphy@google.com>
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>