drivers: serial: nrfx_uarte: Add workaround for FRAMETIMEOUT corner case
When reception is restarted (STARTRX after ENDRX but no STOPRX) it is
possible that FRAMETIMEOUT countdown counter will not be started by
the first received byte if byte was already being transmitted when
STARTRX was called. If that is the only byte then it is expected that
timeout will be triggered but since FRAMETIMEOUT counter is not started
there is no FRAMETIMEOUT event which has short to STOPRX. This
situation will happen in case short buffers are used (< 5 bytes)
because then short ENDRX_STARTRX is not used then.

(cherry picked from commit 94d355a0bdb232479d75db1026f5017ed0fb4dff)

Original-Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>
GitOrigin-RevId: 94d355a0bdb232479d75db1026f5017ed0fb4dff
Cr-Build-Id: 8712248484832760737
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8712248484832760737
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Id3bf4db7336578d188a47ac118ab0d0bd26c76ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6640463
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>