soc/mediatek/mtk_adsp: Use smaller accesses when find()ing in device memory
Recent Python interpreters have started tossing bus errors from this
12-byte string search (the loader is looking for the winstream
descriptor in the live firmware image).  My guess is that there's a
SIMD optimization that's been added that's trying to do e.g. a 16 byte
load, and something in the fabric is kicking that out.  Note that this
is 100% a software change: the same hardware with one version of the
host environment works, and an update breaks it.

But really I have no idea what's happening here, the memory region in
question is documented as system DRAM, the same bus regular process
memory is on (it's just not kernel-utilized memory).  All I know is
that the bus error is introduced with a Python upgrade from 3.8.20 to
3.11.10.

Regardless, it's no great hardship to do the search on 64 bit chunks.

(cherry picked from commit 0f4eeb6380db5734fd234e44dfa2e4d9e8419c8c)

Original-Signed-off-by: Andy Ross <andyross@google.com>
GitOrigin-RevId: 0f4eeb6380db5734fd234e44dfa2e4d9e8419c8c
Cr-Build-Id: 8723006874034296081
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8723006874034296081
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I64409db62f8b48d6acf876f771459a06facabc0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6266723
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>