soc/mediatek/mtk_adsp: Always cache the full SRAM region
I think my original idea with this default MPU setup was that the top
bits of the (fast) SRAM region might be useful for host DMA that
needed better latencies than the (extremely slow) system DRAM
mappings.  So it should be left uncached for safety.

But unfortunately the author[1] of the SOF heap integration for this
platform decided to size the heap dynamically to use most of the SRAM
block (the vectors and a few other bits live at the bottom, but most
of .text is in DRAM).

Needless to say, an uncached heap is sort of a performance disaster.
It worked OK for default copy-only topologies but fell over the moment
we turned on nontrivial processing.

[1] Um... Hi.  Yeah, that's me too.

(cherry picked from commit 4b27b5494fb8d9f342c5e7bdd11aa9c2845b471c)

Original-Signed-off-by: Andy Ross <andyross@google.com>
GitOrigin-RevId: 4b27b5494fb8d9f342c5e7bdd11aa9c2845b471c
Cr-Build-Id: 8723006874034296081
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8723006874034296081
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I4c13188d4c1b601dc06d4ab032763f2c6c8d4fed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6266724
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>